data_research/requirements.txt
core==1.0.1
pandas==3.0.2
python-dotenv==1.2.2
scikit_learn==1.8.0
